requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
idna==3.4
importlib-metadata==7.0.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
preshed==3.0.9
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thinc==8.2.2
threadpoolctl==3.2.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
zipp==3.17.0
